pyproject.toml
[project]
name = "qq-bot-test"
version = "0.1.0"
description = "QQ机器人测试项目"
requires-python = ">=3.8"
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

